pinned-requirements.txt
Jinja2==3.1.4
MarkupSafe==3.0.1
PyYAML==6.0.2
setuptools==75.1.0
